+++ src/bci_build/package.py
@@ -390,12 +390,8 @@ class BaseContainerImage(abc.ABC):
 
     _image_properties: ImageProperties = field(default=_SLE_IMAGE_PROPS)
 
-    #: don't add the packages from :py:attr:`package_list` to the dockerfile or
-    #: kiwi xml
-    _no_default_packages: bool = False
-
     def __post_init__(self) -> None:
-        if not self.package_list and not self._no_default_packages:
+        if not self.package_list:
             raise ValueError(f"No packages were added to {self.pretty_name}.")
         if self.exclusive_arch and Arch.LOCAL in self.exclusive_arch:
             raise ValueError(f"{Arch.LOCAL} must not appear in {self.exclusive_arch=}")
@@ -616,8 +612,6 @@ exit 0
         :command:`zypper in`.
 
         """
-        if self._no_default_packages:
-            return ""
         for pkg in self.package_list:
             if isinstance(pkg, Package) and pkg.pkg_type != PackageType.IMAGE:
                 raise ValueError(
@@ -712,8 +706,6 @@ exit 0
         """The package list as xml elements that are inserted into a kiwi build
         description file.
         """
-        if self._no_default_packages:
-            return ""
 
         def create_pkg_filter_func(
             pkg_type: PackageType,
@@ -913,24 +905,11 @@ exit 0
 
         return ",".join(extra_tags) if extra_tags else None
 
-    async def late_init(self) -> None:
-        """Optional costly initialization function that can be implemented in
-        child-classes.
-
-        This function is called before writing the build recipes in
-        :py:func:`write_files_to_folder` and can be used for long running
-        operations that should not block other operations.
-
-        """
-        pass
-
     async def write_files_to_folder(self, dest: str) -> List[str]:
         """Writes all files required to build this image into the destination folder and
         returns the filenames (not full paths) that were written to the disk.
 
         """
-        await self.late_init()
-
         files = ["_service"]
         tasks = []
 
@@ -1021,9 +1000,6 @@ class LanguageStackContainer(BaseContainerImage):
     #: flag whether the version should be included in the uid
     version_in_uid: bool = True
 
-    #: flag to not include ``$name-$version_label`` in :py:attr:`build_tags`
-    _no_default_version: bool = False
-
     def __post_init__(self) -> None:
         super().__post_init__()
         if not self.version:
@@ -1045,9 +1021,7 @@ class LanguageStackContainer(BaseContainerImage):
     def build_tags(self) -> List[str]:
         tags = []
         for name in [self.name] + self.additional_names:
-            for ver_label in (
-                [self.version_label] if not self._no_default_version else []
-            ) + self.additional_versions:
+            for ver_label in [self.version_label] + self.additional_versions:
                 tags += [f"{self._registry_prefix}/{name}:{ver_label}"] + [
                     f"{self._registry_prefix}/{name}:{ver_label}-%RELEASE%"
                 ]
